clean up demo pages and added a default one
